✨ feat: update package version to 0.6.1 and update zustand library
The commit updates the version of the `multiplatform.one` package from `0.6.0` to `0.6.1`. Additionally, it updates the `zustand` library by changing the file `index.d.ts` and its related types and importing a new version of the `createSimple` function. Finally, the commit also includes the addition of a new function `setDefaultCrossStorage` and exports several types and functions.