✨ feat: Add vertical and horizontal display options for FormProgress component, update progress bar rotation angle
This commit adds two display options to the FormProgress component in packages/components/src/forms/FormProgress/index.tsx. The `vertical` option is added with a name prop of "progress" and spread args. A new `horizontal` option is also added using the same props as the `vertical` option. Additionally, the rotation angle of the progress bar is set to 270 degrees for vertical display.

In another file, packages/components/src/organize/SimpleTabs/SimpleTabs.stories.tsx, the default route is changed to "contact" and full screen is enabled for the `horizontalTabs` option.