✨ feat: Add closeOnHoverOut functionality to SimplePopover
This commit adds the `closeOnHoverOut` prop to the `SimplePopover` component in order to enable closing the popover when hovering out. The `closeOnHoverOut` prop is utilized in the `SimplePopover` component's implementation, specifically in the `onHoverOut` event handler. When `triggerOnHover` or `closeOnHoverOut` is true, the `setOpen(false)` function is called to close the popover.