feat(api): Add 'getFoo' and 'postFoo' endpoints to RocketController
This commit adds two new endpoints, 'getFoo' and 'postFoo', to the RocketController in the api/src/modules/rocket/rocket.controller.ts file. The 'getFoo' endpoint returns a Promise of type FooBar with the properties hello and world set to 'abc' and '123' respectively. The 'postFoo' endpoint accepts a request body of type FooBar and returns the same object. These endpoints are implemented using the corresponding methods in the RocketController class.

Changes also include importing the FooBar type from './rocket.resolver' in the rocket.controller.ts file, which is used as the return type for the 'getFoo' endpoint.