fix(service): Remove todo from the list when found
This commit fixes an issue in the `TodoService` where a todo is removed from the list when it is found. It modifies the `splice` method to delete the todo at the specified index and returns `true` upon successful removal. If the todo is not found, an error is thrown with the corresponding message.