✨ feat: Add arrow option to SimpleTooltip component
This commit adds an arrow option to the SimpleTooltip component in order to control whether or not the tooltip has an arrow. The arrow state is managed using React hooks and can be toggled via a Switch component. The arrow is conditionally rendered within the Tooltip.Content component if the arrow option is set to true.